Changed options term from 'Portfolio Category Slug' to 'Portfolio Permastruct Slug'.
Changed the call to flush_rewrite_rules to call it from a register_activation_hook function and after a call to a new function that registers the post type.  This new function also is called during the INIT.  Also added calling flush_rewrite_rules from the deactivation hook.

Fixed typo on one of the var names for the new mobile styling options.