Fixed minor activation issue and correctly word-wrapped comment.
Discovered a minor issue with the plugin registration and code
template. When the plugin would activate, the option items would be
initialized with default values. These values are not accounted in
the code template, which means that the default values would
initialize the snippet rendering, this was not intended!

This issue has been addressed and fixed. The default values when
the plugin is initialized is null on both fields. These are also
the default values for the "get_option"-function calls within the
code template, hence the code template works as intended.

Fixed minor comment wrapping issue, according to the 80 character
line ending convention.